+++ src/main/java/RenovatioMod/renovatio/attribute/ModAttributes.java
@@ -1,0 +1,25 @@
+package RenovatioMod.renovatio.attribute;
+import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.ClampedEntityAttribute;
+import net.minecraft.entity.attribute.EntityAttribute;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
+import net.minecraft.util.Identifier;
+import RenovatioMod.renovatio.Renovatio;
+
+public class ModAttributes {
+    public static final EntityAttribute GENERIC_MAX_TOUGHNESS = new ClampedEntityAttribute(
+            "attribute.name.generic.max_toughness",
+            20.0, // Default value
+            0.0,   // Min value
+            1024.0 // Max value
+    ).setTracked(true);
+
+    public static void registerAttributes() {
+        Registry.register(Registries.ATTRIBUTE, new Identifier(Renovatio.MOD_ID, "generic.max_toughness"), GENERIC_MAX_TOUGHNESS);
+        // Add the attribute to players by default
+        FabricDefaultAttributeRegistry.register(EntityType.PLAYER, LivingEntity.createLivingAttributes().add(GENERIC_MAX_TOUGHNESS));
+    }
+}